Add files via upload
GROUP Members:
19SW33
19SW48
19SW50
19SW122

PROJECT DESCRIPTION:
The main objective of the Tourism Management System is to manage the details of Customer, Hotel Booking, Cancellation and Tourism places. It manages all the information about Users, Hotel, Packages etc. The project is totally built at administrative end and thus only the administrator is guaranteed the access to the backend database. The purpose of this project is to build an application program to reduce the manual work for managing Tourists, Booking, Places etc. •MySQL 
Programming language: JAVA